clang-tidy: fix tautological compare
eSELData[2] is a char and 0xDD exceeds signed char range, making the
comparison always false. Compare as unsigned by casting to uint8_t
to ensure correct semantics and silence the warning.

Fixed below error
'''
../oemhandler.cpp:310:25: error: result of comparison of constant 221 with expression of type 'value_type' (aka 'char') is always false [-Werror,-Wtautological-constant-out-of-range-compare]
  310 |         if (eSELData[2] == occMetricsType)
'''

+++ oemhandler.cpp
@@ -307,7 +307,8 @@ ipmi_ret_t ipmi_ibm_oem_partial_esel(
 
         // If the eSEL record type is OCC metrics, then create the OCC log
         // entry.
-        if (eSELData[2] == occMetricsType)
+        if (static_cast<uint8_t>(eSELData[2]) ==
+            static_cast<uint8_t>(occMetricsType))
         {
             createOCCLogEntry(eSELData);
         }
